Split out wget API into seperate header and JS implementation
This should be code size saving for apps that use Browser APIs
but not the wget ones
