+++ src/modules.js
@@ -86,6 +86,7 @@ var LibraryManager = {
       libraries.push('library_strings.js');
     } else {
       libraries.push('library_browser.js');
+      libraries.push('library_wget.js');
     }
 
     if (USE_PTHREADS) { // TODO: Currently WebGL proxying makes pthreads library depend on WebGL.
