+++ system/include/emscripten/wget.h
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2012 The Emscripten Authors.  All rights reserved.
+ * Emscripten is available under two separate licenses, the MIT license and the
+ * University of Illinois/NCSA Open Source License.  Both these licenses can be
+ * found in the LICENSE file.
+ */
+
+#pragma once
+
+#include "em_types.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+// wget
+
+void emscripten_async_wget(const char* url, const char* file, em_str_callback_func onload, em_str_callback_func onerror);
+
+typedef void (*em_async_wget_onload_func)(void*, void*, int);
+void emscripten_async_wget_data(const char* url, void *arg, em_async_wget_onload_func onload, em_arg_callback_func onerror);
+
+typedef void (*em_async_wget2_onload_func)(unsigned, void*, const char*);
+typedef void (*em_async_wget2_onstatus_func)(unsigned, void*, int);
+
+int emscripten_async_wget2(const char* url, const char* file,  const char* requesttype, const char* param, void *arg, em_async_wget2_onload_func onload, em_async_wget2_onstatus_func onerror, em_async_wget2_onstatus_func onprogress);
+
+typedef void (*em_async_wget2_data_onload_func)(unsigned, void*, void*, unsigned);
+typedef void (*em_async_wget2_data_onerror_func)(unsigned, void*, int, const char*);
+typedef void (*em_async_wget2_data_onprogress_func)(unsigned, void*, int, int);
+
+int emscripten_async_wget2_data(const char* url, const char* requesttype, const char* param, void *arg, int free, em_async_wget2_data_onload_func onload, em_async_wget2_data_onerror_func onerror, em_async_wget2_data_onprogress_func onprogress);
+
+void emscripten_async_wget2_abort(int handle);
+
+// wget "sync"
+
+void emscripten_wget(const char* url, const char* file);
+void emscripten_wget_data(const char* url, void** pbuffer, int* pnum, int *perror);
+
+#ifdef __cplusplus
+}
+#endif
